Organize crypto tests
